refactor: 일관성을 위해 엔티티 파일 이름 수정
-theaterMovie 엔티티를 theater-movie로 변경
- 그에 따라 import 문 업데이트
