+++ src/theater-movies/entities/theater-movie.entity.ts
